+++ src/Backupper/MysqlBackupper.php
@@ -11,24 +11,23 @@ class MysqlBackupper extends AbstractBackupper
     #[\Override]
     public function buildCommandLine(): CommandLine
     {
-        $dbParams = $this->connection->getParams();
         $command = new CommandLine($this->binary);
 
-        if (isset($dbParams['host'])) {
-            $command->addArg('-h', $dbParams['host']);
+        if ($host = $this->databaseDsn->getHost()) {
+            $command->addArg('-h', $host);
         }
-        if (isset($dbParams['user'])) {
-            $command->addArg('-u', $dbParams['user']);
+        if ($user = $this->databaseDsn->getUser()) {
+            $command->addArg('-u', $user);
         }
-        if (isset($dbParams['port'])) {
-            $command->addArg('-P', $dbParams['port']);
+        if ($port = $this->databaseDsn->getPort()) {
+            $command->addArg('-P', $port);
         }
-        if (isset($dbParams['password'])) {
-            $command->addArg('-p' . $dbParams['password']);
+        if ($password = $this->databaseDsn->getPassword()) {
+            $command->addArg('-p' . $password);
         }
 
         foreach ($this->excludedTables as $table) {
-            $command->addArg('--ignore-table', $dbParams['dbname'] . '.' . $table);
+            $command->addArg('--ignore-table', $this->databaseDsn->getDatabase() . '.' . $table);
         }
 
         if ($this->verbose) {
@@ -41,7 +40,7 @@ class MysqlBackupper extends AbstractBackupper
             $command->addArg('-r', $this->destination);
         }
 
-        $command->addArg($dbParams['dbname']);
+        $command->addArg($this->databaseDsn->getDatabase());
 
         return $command;
     }