+++ src/Restorer/AbstractRestorer.php
@@ -4,10 +4,11 @@ declare(strict_types=1);
 
 namespace MakinaCorpus\DbToolsBundle\Restorer;
 
-use Doctrine\DBAL\Connection;
 use MakinaCorpus\DbToolsBundle\Helper\Output\NullOutput;
 use MakinaCorpus\DbToolsBundle\Helper\Process\CommandLine;
 use MakinaCorpus\DbToolsBundle\Helper\Process\ProcessTrait;
+use MakinaCorpus\QueryBuilder\DatabaseSession;
+use MakinaCorpus\QueryBuilder\Dsn;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\NullLogger;
 use Symfony\Component\Process\Process;
@@ -27,7 +28,8 @@ abstract class AbstractRestorer implements LoggerAwareInterface
 
     public function __construct(
         protected string $binary,
-        protected Connection $connection,
+        protected DatabaseSession $databaseSession,
+        protected Dsn $databaseDsn,
         ?string $defaultOptions = null,
     ) {
         $this->defaultOptions = $defaultOptions ?? $this->getBuiltinDefaultOptions();